rearrange files and functions, no change to functionality
