+++ src/matrices_tuples/tuples_create.c
@@ -1,0 +1,23 @@
+#include "minirt.h"
+
+t_tuple	create_vector(float x, float y, float z)
+{
+	t_tuple	new_vector;
+
+	new_vector.x = x;
+	new_vector.y = y;
+	new_vector.z = z;
+	new_vector.w = 0;
+	return (new_vector);
+}
+
+t_tuple	create_point(float x, float y, float z)
+{
+	t_tuple	new_point;
+
+	new_point.x = x;
+	new_point.y = y;
+	new_point.z = z;
+	new_point.w = 1;
+	return (new_point);
+}
