+++ src/utils/is_equal.c
